requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
loguru==0.7.3
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
supabase
pymupdf4llm 
sentence-transformers
chromadb
google-generativeai
langchain-google-genai
langchain
langgraph
langchain-chroma
tqdm
python-dotenv
python-multipart